% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10am.charlson-package.R
\docType{package}
\name{icd10am.charlson-package}
\alias{icd10am.charlson}
\alias{icd10am.charlson-package}
\title{icd10am.charlson: Charlson Comorbidity Index for ICD-10-AM Coded Data}
\description{
Applies the Charlson Comorbidity Index to Australian admitted patient care data coded using ICD-10-AM. Supports both wide and long input formats and returns Quan and Sundararajan weighted scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/healthpolicy/icd10am.charlson}
  \item Report bugs at \url{https://github.com/healthpolicy/icd10am.charlson/issues}
}

}
\author{
\strong{Maintainer}: Jim Pearse \email{ejpearse@gmail.com}

}
\keyword{internal}
